refactor(utils): moved /pkg/synchronizer/utils folder to /pkg/utils
